// lib/types/cdnTweet.d.ts
export default interface CDNTweet {
    __typename: string;
    lang: string;
    favorite_count: number;
    possibly_sensitive?: boolean;
    created_at: string;
    display_text_range?: number[];
    entities?: IEntities;
    id_str: string;
    text: string;
    user: IUser;
    edit_control: IEdit_control;
    mediaDetails?: IMediaDetailsItem[];
    photos?: IPhotosItem[];
    conversation_count: number;
    news_action_type: string;
    isEdited: boolean;
    isStaleEdit: boolean;
    place?: IPlace;
    quoted_tweet?: IQuoted_tweet;
    in_reply_to_screen_name?: string;
    in_reply_to_status_id_str?: string;
    in_reply_to_user_id_str?: string;
    self_thread?: ISelf_thread;
    parent?: IParent;
    video?: IVideo;
    covid19_hub_url?: string;
    card?: ICard;
    conversation_control?: IConversation_control;
}
interface IEntities {
    hashtags: IHashtagsItem[];
    urls: IUrlsItem[];
    user_mentions: IUserMentionsItem[];
    symbols: ISymbolsItem[];
    media?: IMediaItem[];
}
interface IUrlsItem {
    display_url: string;
    expanded_url: string;
    indices: number[];
    url: string;
}
interface IMediaItem {
    display_url: string;
    expanded_url: string;
    indices: number[];
    url: string;
}
interface IUser {
    id_str: string;
    name: string;
    profile_image_url_https: string;
    screen_name: string;
    verified: boolean;
    is_blue_verified: boolean;
    verified_type?: string;
    highlighted_label?: IHighlighted_label;
}
interface IEdit_control {
    edit_tweet_ids: string[];
    editable_until_msecs: string;
    is_edit_eligible: boolean;
    edits_remaining: string;
}
interface IMediaDetailsItem {
    display_url: string;
    expanded_url: string;
    ext_media_availability: IExt_media_availability;
    indices: number[];
    media_url_https: string;
    original_info: IOriginal_info;
    sizes: ISizes;
    type: string;
    url: string;
    additional_media_info?: IAdditional_media_info;
    ext_media_stats?: IExt_media_stats;
    video_info?: IVideo_info;
    ext_alt_text?: string;
}
interface IExt_media_availability {
    status: string;
    reason?: string;
}
interface IOriginal_info {
    height: number;
    width: number;
    focus_rects?: IFocusRectsItem[];
}
interface IFocusRectsItem {
    x: number;
    y: number;
    w: number;
    h: number;
}
interface ISizes {
    large: ILarge;
    medium: IMedium;
    small: ISmall;
    thumb: IThumb;
}
interface ILarge {
    h: number;
    resize: string;
    w: number;
}
interface IMedium {
    h: number;
    resize: string;
    w: number;
}
interface ISmall {
    h: number;
    resize: string;
    w: number;
}
interface IThumb {
    h: number;
    resize: string;
    w: number;
}
interface IPhotosItem {
    backgroundColor: IBackgroundColor;
    cropCandidates: ICropCandidatesItem[];
    expandedUrl: string;
    url: string;
    width: number;
    height: number;
    accessibilityLabel?: string;
}
interface IBackgroundColor {
    red: number;
    green: number;
    blue: number;
}
interface ICropCandidatesItem {
    x: number;
    y: number;
    w: number;
    h: number;
}
interface IPlace {
    full_name: string;
    id: string;
}
interface IQuoted_tweet {
    in_reply_to_screen_name?: string;
    in_reply_to_status_id_str?: string;
    in_reply_to_user_id_str?: string;
    lang: string;
    place?: IPlace;
    reply_count: number;
    retweet_count: number;
    favorite_count: number;
    created_at: string;
    display_text_range: number[];
    entities: IEntities;
    id_str: string;
    text: string;
    user: IUser;
    edit_control?: IEdit_control;
    isEdited?: boolean;
    isStaleEdit?: boolean;
    possibly_sensitive?: boolean;
    mediaDetails?: IMediaDetailsItem[];
    photos?: IPhotosItem[];
    self_thread?: ISelf_thread;
    video?: IVideo;
    conversation_control?: IConversation_control;
}
interface IUserMentionsItem {
    id_str: string;
    indices: number[];
    name: string;
    screen_name: string;
}
interface ISelf_thread {
    id_str: string;
}
interface IHashtagsItem {
    indices: number[];
    text: string;
}
interface IParent {
    lang: string;
    reply_count: number;
    retweet_count: number;
    favorite_count: number;
    self_thread?: ISelf_thread;
    possibly_sensitive?: boolean;
    created_at: string;
    display_text_range: number[];
    entities: IEntities;
    id_str: string;
    text: string;
    user: IUser;
    edit_control: IEdit_control;
    mediaDetails?: IMediaDetailsItem[];
    photos?: IPhotosItem[];
    isEdited: boolean;
    isStaleEdit: boolean;
    in_reply_to_screen_name?: string;
    in_reply_to_status_id_str?: string;
    in_reply_to_user_id_str?: string;
    place?: IPlace;
    video?: IVideo;
    conversation_control?: IConversation_control;
    quoted_tweet?: IQuoted_tweet;
    card?: ICard;
}
interface IAdditional_media_info {
    description?: string;
    embeddable?: boolean;
    title?: string;
}
interface IExt_media_stats {
    view_count: number;
}
interface IVideo_info {
    aspect_ratio: number[];
    duration_millis?: number;
    variants: IVariantsItem[];
}
interface IVariantsItem {
    bitrate?: number;
    content_type?: string;
    url?: string;
    type?: string;
    src?: string;
}
interface IVideo {
    aspectRatio: number[];
    contentType?: string;
    durationMs?: number;
    mediaAvailability?: IMediaAvailability;
    poster: string;
    variants?: IVariantsItem[];
    videoId?: IVideoId;
    viewCount?: number;
}
interface IMediaAvailability {
    status: string;
    reason?: string;
}
interface IVideoId {
    type: string;
    id: string;
}
interface ICard {
    card_platform: ICard_platform;
    name: string;
    url: string;
    binding_values: IBinding_values;
}
interface ICard_platform {
    platform: IPlatform;
}
// ...
interface IHighlighted_label {
    description: string;
    badge: IBadge;
    url?: IUrl;
    user_label_type?: string;
    user_label_display_type?: string;
}
// ...
interface IConversation_control {
    policy: string;
}
// ...
interface ISymbolsItem {
    indices: number[];
    text: string;
}